Refactor: move engine handlers to subfolder and move tests to standalone files
