+++ openraft/src/engine/handler/replication_handler/mod.rs
@@ -21,6 +21,9 @@ use crate::NodeId;
 use crate::RaftState;
 use crate::ServerState;
 
+#[cfg(test)] mod append_membership_test;
+#[cfg(test)] mod update_matching_test;
+
 /// Handle replication operations.
 ///
 /// - Writing local log store;
@@ -401,362 +404,6 @@ where
             config: self.config,
             state: self.state,
             output: self.output,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-
-    mod update_matching_test {
-
-        use std::sync::Arc;
-
-        use maplit::btreeset;
-        use pretty_assertions::assert_eq;
-        use tokio::time::Instant;
-
-        use crate::engine::Command;
-        use crate::engine::Engine;
-        use crate::progress::Inflight;
-        use crate::progress::Progress;
-        use crate::raft_state::LogStateReader;
-        use crate::testing::log_id;
-        use crate::utime::UTime;
-        use crate::EffectiveMembership;
-        use crate::Membership;
-        use crate::MembershipState;
-        use crate::Vote;
-
-        fn m01() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {0,1}], None)
-        }
-
-        fn m123() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {1,2,3}], None)
-        }
-
-        fn eng() -> Engine<u64, ()> {
-            let mut eng = Engine::default();
-            eng.state.enable_validate = false; // Disable validation for incomplete state
-
-            eng.config.id = 2;
-            eng.state.vote = UTime::new(Instant::now(), Vote::new_committed(2, 1));
-            eng.state.membership_state = MembershipState::new(
-                Arc::new(EffectiveMembership::new(Some(log_id(1, 1)), m01())),
-                Arc::new(EffectiveMembership::new(Some(log_id(2, 3)), m123())),
-            );
-
-            eng
-        }
-
-        #[test]
-        fn test_update_matching_no_leader() -> anyhow::Result<()> {
-            // There is no leader, it should panic.
-
-            let res = std::panic::catch_unwind(move || {
-                let mut eng = eng();
-                eng.replication_handler().update_matching(3, 0, Some(log_id(1, 2)));
-            });
-            tracing::info!("res: {:?}", res);
-            assert!(res.is_err());
-
-            Ok(())
-        }
-
-        #[test]
-        fn test_update_matching() -> anyhow::Result<()> {
-            let mut eng = eng();
-            eng.vote_handler().become_leading();
-
-            let mut rh = eng.replication_handler();
-            let inflight_id_1 = {
-                let prog_entry = rh.leader.progress.get_mut(&1).unwrap();
-                prog_entry.inflight = Inflight::logs(Some(log_id(2, 3)), Some(log_id(2, 4)));
-                prog_entry.inflight.get_id().unwrap()
-            };
-            let inflight_id_2 = {
-                let prog_entry = rh.leader.progress.get_mut(&2).unwrap();
-                prog_entry.inflight = Inflight::logs(Some(log_id(1, 0)), Some(log_id(2, 4)));
-                prog_entry.inflight.get_id().unwrap()
-            };
-            let inflight_id_3 = {
-                let prog_entry = rh.leader.progress.get_mut(&3).unwrap();
-                prog_entry.inflight = Inflight::logs(Some(log_id(1, 1)), Some(log_id(2, 4)));
-                prog_entry.inflight.get_id().unwrap()
-            };
-
-            // progress: None, None, (1,2)
-            {
-                rh.update_matching(3, inflight_id_3, Some(log_id(1, 2)));
-                assert_eq!(None, rh.state.committed());
-                assert_eq!(
-                    vec![
-                        //
-                        Command::UpdateProgressMetrics {
-                            target: 3,
-                            matching: log_id(1, 2),
-                        },
-                    ],
-                    rh.output.commands
-                );
-            }
-
-            // progress: None, (2,1), (1,2); quorum-ed: (1,2), not at leader vote, not committed
-            {
-                rh.output.commands = vec![];
-                rh.update_matching(2, inflight_id_2, Some(log_id(2, 1)));
-                assert_eq!(None, rh.state.committed());
-                assert_eq!(0, rh.output.commands.len());
-            }
-
-            // progress: None, (2,1), (2,3); committed: (2,1)
-            {
-                rh.output.commands = vec![];
-                rh.update_matching(3, inflight_id_3, Some(log_id(2, 3)));
-                assert_eq!(Some(&log_id(2, 1)), rh.state.committed());
-                assert_eq!(
-                    vec![
-                        Command::UpdateProgressMetrics {
-                            target: 3,
-                            matching: log_id(2, 3),
-                        },
-                        Command::ReplicateCommitted {
-                            committed: Some(log_id(2, 1))
-                        },
-                        Command::LeaderCommit {
-                            already_committed: None,
-                            upto: log_id(2, 1)
-                        }
-                    ],
-                    rh.output.commands
-                );
-            }
-
-            // progress: (2,4), (2,1), (2,3); committed: (1,3)
-            {
-                rh.output.commands = vec![];
-                rh.update_matching(1, inflight_id_1, Some(log_id(2, 4)));
-                assert_eq!(Some(&log_id(2, 3)), rh.state.committed());
-                assert_eq!(
-                    vec![
-                        Command::UpdateProgressMetrics {
-                            target: 1,
-                            matching: log_id(2, 4),
-                        },
-                        Command::ReplicateCommitted {
-                            committed: Some(log_id(2, 3))
-                        },
-                        Command::LeaderCommit {
-                            already_committed: Some(log_id(2, 1)),
-                            upto: log_id(2, 3)
-                        }
-                    ],
-                    rh.output.commands
-                );
-            }
-
-            Ok(())
-        }
-    }
-    mod append_membership_test {
-
-        use std::sync::Arc;
-
-        use maplit::btreeset;
-        use tokio::time::Instant;
-
-        use crate::core::ServerState;
-        use crate::engine::Command;
-        use crate::engine::Engine;
-        use crate::engine::LogIdList;
-        use crate::progress::entry::ProgressEntry;
-        use crate::progress::Inflight;
-        use crate::progress::Progress;
-        use crate::utime::UTime;
-        use crate::CommittedLeaderId;
-        use crate::EffectiveMembership;
-        use crate::LogId;
-        use crate::Membership;
-        use crate::MembershipState;
-        use crate::MetricsChangeFlags;
-        use crate::Vote;
-
-        crate::declare_raft_types!(
-            pub(crate) Foo: D=(), R=(), NodeId=u64, Node=()
-        );
-
-        use crate::testing::log_id;
-
-        fn m01() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {0,1}], None)
-        }
-
-        fn m23() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {2,3}], None)
-        }
-
-        fn m23_45() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {2,3}], Some(btreeset! {4,5}))
-        }
-
-        fn m34() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {3,4}], None)
-        }
-
-        fn m4_356() -> Membership<u64, ()> {
-            Membership::<u64, ()>::new(vec![btreeset! {4}], Some(btreeset! {3,5,6}))
-        }
-
-        fn eng() -> Engine<u64, ()> {
-            let mut eng = Engine::default();
-            eng.config.id = 2;
-            eng.state.membership_state = MembershipState::new(
-                Arc::new(EffectiveMembership::new(Some(log_id(1, 1)), m01())),
-                Arc::new(EffectiveMembership::new(Some(log_id(2, 3)), m23())),
-            );
-            eng.state.vote = UTime::new(Instant::now(), Vote::new_committed(2, 2));
-            eng.state.server_state = eng.calc_server_state();
-            eng
-        }
-
-        #[test]
-        fn test_leader_append_membership_for_leader() -> anyhow::Result<()> {
-            let mut eng = eng();
-            eng.state.server_state = ServerState::Leader;
-            // Make it a real leader: voted for itself and vote is committed.
-            eng.state.vote = UTime::new(Instant::now(), Vote::new_committed(2, 2));
-            eng.vote_handler().become_leading();
-
-            eng.replication_handler().append_membership(&log_id(3, 4), &m34());
-
-            assert_eq!(
-                MembershipState::new(
-                    Arc::new(EffectiveMembership::new(Some(log_id(2, 3)), m23())),
-                    Arc::new(EffectiveMembership::new(Some(log_id(3, 4)), m34()))
-                ),
-                eng.state.membership_state
-            );
-            assert_eq!(
-                ServerState::Leader,
-                eng.state.server_state,
-                "Leader wont be affected by membership change"
-            );
-
-            assert_eq!(
-                MetricsChangeFlags {
-                    replication: true,
-                    local_data: false,
-                    cluster: true,
-                },
-                eng.output.metrics_flags
-            );
-
-            assert_eq!(
-                vec![
-                    //
-                    Command::UpdateMembership {
-                        membership: Arc::new(EffectiveMembership::new(Some(log_id(3, 4)), m34())),
-                    },
-                    Command::RebuildReplicationStreams {
-                        targets: vec![(3, ProgressEntry::empty(0)), (4, ProgressEntry::empty(0))], /* node-2 is leader,
-                                                                                                    * won't be removed */
-                    }
-                ],
-                eng.output.commands
-            );
-
-            assert!(
-                eng.internal_server_state.leading().unwrap().progress.get(&4).matching.is_none(),
-                "exists, but it is a None"
-            );
-
-            Ok(())
-        }
-
-        #[test]
-        fn test_leader_append_membership_update_learner_process() -> anyhow::Result<()> {
-            // When updating membership, voter progreess should inherit from learner progress, and
-            // learner process should inherit from voter process. If voter changes to
-            // learner or vice versa.
-
-            let mut eng = eng();
-            eng.state.log_ids =
-                LogIdList::new([LogId::new(CommittedLeaderId::new(0, 0), 0), log_id(1, 1), log_id(5, 10)]);
-
-            eng.state.server_state = ServerState::Leader;
-            // Make it a real leader: voted for itself and vote is committed.
-            eng.state.vote = UTime::new(Instant::now(), Vote::new_committed(2, 2));
-            eng.state
-                .membership_state
-                .set_effective(Arc::new(EffectiveMembership::new(Some(log_id(2, 3)), m23_45())));
-            eng.vote_handler().become_leading();
-
-            if let Some(l) = &mut eng.internal_server_state.leading_mut() {
-                assert_eq!(&ProgressEntry::empty(11), l.progress.get(&4));
-                assert_eq!(&ProgressEntry::empty(11), l.progress.get(&5));
-
-                let p = ProgressEntry::new(Some(log_id(1, 4)));
-                let _ = l.progress.update(&4, p);
-                assert_eq!(&p, l.progress.get(&4));
-
-                let p = ProgressEntry::new(Some(log_id(1, 5)));
-                let _ = l.progress.update(&5, p);
-                assert_eq!(&p, l.progress.get(&5));
-
-                let p = ProgressEntry::new(Some(log_id(1, 3)));
-                let _ = l.progress.update(&3, p);
-                assert_eq!(&p, l.progress.get(&3));
-            } else {
-                unreachable!("leader should not be None");
-            }
-
-            eng.replication_handler().append_membership(&log_id(3, 4), &m4_356());
-
-            assert_eq!(
-                MembershipState::new(
-                    Arc::new(EffectiveMembership::new(Some(log_id(2, 3)), m23_45())),
-                    Arc::new(EffectiveMembership::new(Some(log_id(3, 4)), m4_356()))
-                ),
-                eng.state.membership_state
-            );
-
-            if let Some(l) = &mut eng.internal_server_state.leading_mut() {
-                assert_eq!(
-                    &ProgressEntry::new(Some(log_id(1, 4)))
-                        .with_inflight(Inflight::logs(Some(log_id(1, 4)), Some(log_id(5, 10))).with_id(1))
-                        .with_curr_inflight_id(1),
-                    l.progress.get(&4),
-                    "learner-4 progress should be transferred to voter progress"
-                );
-
-                assert_eq!(
-                    &ProgressEntry::new(Some(log_id(1, 3)))
-                        .with_inflight(Inflight::logs(Some(log_id(1, 3)), Some(log_id(5, 10))).with_id(1))
-                        .with_curr_inflight_id(1),
-                    l.progress.get(&3),
-                    "voter-3 progress should be transferred to learner progress"
-                );
-
-                assert_eq!(
-                    &ProgressEntry::new(Some(log_id(1, 5)))
-                        .with_inflight(Inflight::logs(Some(log_id(1, 5)), Some(log_id(5, 10))).with_id(1))
-                        .with_curr_inflight_id(1),
-                    l.progress.get(&5),
-                    "learner-5 has previous value"
-                );
-
-                assert_eq!(
-                    &ProgressEntry::empty(11)
-                        .with_inflight(Inflight::logs(None, Some(log_id(5, 10))).with_id(1))
-                        .with_curr_inflight_id(1),
-                    l.progress.get(&6)
-                );
-            } else {
-                unreachable!("leader should not be None");
-            }
-
-            Ok(())
         }
     }
 }
