+++ openraft/src/engine/handler/server_state_handler/update_server_state_test.rs
@@ -1,0 +1,67 @@
+use std::sync::Arc;
+
+use maplit::btreeset;
+use pretty_assertions::assert_eq;
+use tokio::time::Instant;
+
+use crate::engine::Command;
+use crate::engine::Engine;
+use crate::testing::log_id;
+use crate::utime::UTime;
+use crate::EffectiveMembership;
+use crate::Membership;
+use crate::MembershipState;
+use crate::ServerState;
+use crate::Vote;
+
+fn m01() -> Membership<u64, ()> {
+    Membership::<u64, ()>::new(vec![btreeset! {0,1}], None)
+}
+
+fn m123() -> Membership<u64, ()> {
+    Membership::<u64, ()>::new(vec![btreeset! {1,2,3}], None)
+}
+
+fn eng() -> Engine<u64, ()> {
+    let mut eng = Engine::default();
+    eng.state.enable_validate = false; // Disable validation for incomplete state
+
+    eng.config.id = 2;
+    eng.state.vote = UTime::new(Instant::now(), Vote::new_committed(2, 2));
+    eng.state.membership_state = MembershipState::new(
+        Arc::new(EffectiveMembership::new(Some(log_id(1, 1)), m01())),
+        Arc::new(EffectiveMembership::new(Some(log_id(2, 3)), m123())),
+    );
+    eng.state.server_state = eng.state.calc_server_state(&eng.config.id);
+
+    eng
+}
+#[test]
+fn test_update_server_state_if_changed() -> anyhow::Result<()> {
+    //
+    let mut eng = eng();
+    let mut ssh = eng.server_state_handler();
+
+    // Leader become follower
+    {
+        assert_eq!(ServerState::Leader, ssh.state.server_state);
+
+        ssh.output.commands = vec![];
+        ssh.state.vote = UTime::new(Instant::now(), Vote::new(2, 100));
+        ssh.update_server_state_if_changed();
+
+        assert_eq!(ServerState::Follower, ssh.state.server_state);
+        assert_eq!(
+            vec![
+                //
+                Command::QuitLeader,
+            ],
+            ssh.output.commands
+        );
+    }
+
+    // TODO(3): add more test,
+    //          after migrating to the no-step-down leader:
+    //          A leader keeps working after it is removed from the voters.
+    Ok(())
+}
